requirements.txt
streamlit==1.32.0
groq==0.5.0
pymupdf==1.23.26
python-dotenv==1.0.1
